+++ src/main/java/net/scilingo/se452/banking/interfaces/IAccount.java
@@ -8,12 +8,12 @@ import net.scilingo.se452.banking.Transaction;
 public interface IAccount {
 	
 	public Integer getId();
-	public Integer getBalance();
+	public Double getBalance();
 	public AccountType getAccountType();
 	public Set<Transaction> getTransactions();
 	
 	public void setId(Integer id);
-	public void setBalance(Integer balance);
+	public void setBalance(Double balance);
 	public void setAccountType(AccountType accountType);
 	public void setTransactions(Set<Transaction> transactions);
